requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# LLM
anthropic>=0.40.0

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio>=0.23.0
black==23.12.0
flake8==6.1.0
